+++ test/com/amplitude/api/SessionTest.java
@@ -17,6 +17,7 @@ import okhttp3.mockwebserver.RecordedRequest;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(RobolectricTestRunner.class)
@@ -884,12 +885,8 @@ public class SessionTest extends BaseTest {
     public void testOnPauseFlushEvents() throws JSONException {
         long timestamp = System.currentTimeMillis();
         long [] timestamps = {
-                timestamp,
-                timestamp + 1,
-                timestamp + 2,
-                timestamp + 3,
-                timestamp + 4,
-                timestamp + 5,
+            timestamp, timestamp + 1, timestamp + 2,
+            timestamp + 3, timestamp + 4, timestamp + 5,
         };
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
         AmplitudeCallbacks callBacks = new AmplitudeCallbacksWithTime(amplitude, timestamps);
@@ -912,5 +909,33 @@ public class SessionTest extends BaseTest {
         // verify that events have been cleared from client
         looper.runOneTask();
         assertEquals(getUnsentEventCount(), 0);
+    }
+
+    @Test
+    public void testOnPauseFlushEventsDisabled() throws JSONException {
+        long timestamp = System.currentTimeMillis();
+        long [] timestamps = {
+            timestamp, timestamp + 1, timestamp + 2,
+            timestamp + 3, timestamp + 4, timestamp + 5,
+        };
+        amplitude.setFlushEventsOnClose(false);
+        ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
+        AmplitudeCallbacks callBacks = new AmplitudeCallbacksWithTime(amplitude, timestamps);
+        Robolectric.getForegroundThreadScheduler().advanceTo(1);
+
+        // log an event, should not be uploaded
+        amplitude.logEventAsync("testEvent", null, null, null, null, timestamps[0], false);
+        looper.runOneTask();
+        looper.runOneTask();
+        assertEquals(getUnsentEventCount(), 1);
+
+        // force client into background and verify no flushing of events
+        callBacks.onActivityPaused(null);
+        looper.runOneTask();  // run the update server
+        RecordedRequest request = runRequest(amplitude);
+
+        // flushing disabled, so no request should be sent
+        assertNull(request);
+        assertEquals(getUnsentEventCount(), 1);
     }
 }