+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-* Add flushing of unsent events on app close/minimize. This only works if you call `Amplitude.getInstance().enableForegroundTracking(getApplication());`, which is recommended in the README by default for Setup.
+* Add automatic flushing of unsent events on app close/minimize (through the Activity Lifecycle `onPause` callback). This only works if you call `Amplitude.getInstance().enableForegroundTracking(getApplication());`, which is recommended in the README by default for Setup. To disable you can call `Amplitude.getInstance().setFlushEventsOnClose(false);`
 
 ## 2.8.0 (June 29, 2016)